Complete overhaul. See commit details.
A law based approach to writing expectations, where classes metaphorically stand in for real world legal equivalents.

Instead of writing 'expectations', a developer writes _laws_ about his/her objects, which are established by the legislature and put to trial. For each law the subject breaks according to the legislation, a punishment is doled out (in our case, an exception is raised).

In order to resolve trial's using the expect(...) syntax where the client is not expected to call -resolve manually, a resolution is performed when any new law is added by the legislature.

The preface class is added but not yet made use of for providing `-to` and `-notTo`.